refactor(web): move limitsFormat into @t3tools/shared (#7420)
